rename ViewBallot -> ViewCart
